+++ lib/src/theme.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class MapChartTheme {
+  MapChartTheme(
+      {this.backgroundColor, required this.foregoundColor, this.contourColor});
+
+  final Color? backgroundColor;
+  final Color foregoundColor;
+  final Color? contourColor;
+}