+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## 0.0.1
+## [0.1.0] - May 2, 2021
+
+* GeoJSON loader
+  * Polygon and MultiPolygon
+* Multi resolution
+  * Geometry simplification in background
+* Geometry highlight on mouse hover
+
+## [0.0.1] - April 25, 2021
 
 * Creating the project